% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{calc_ept_score}
\alias{calc_ept_score}
\title{EPT's score calculation}
\usage{
calc_ept_score(data, blockNum = max(data$block), algoNum = 1)
}
\arguments{
\item{data}{The processed data, with columns:
\itemize{
  \item \code{sid} (Participant's unique ID)
  \item \code{error} (Indicator - 1 is error, 0 is correct)
  \item \code{rt} (Reaction time)
  \item \code{prime_cat} (The primes category)
  \item \code{target_cat} (pos / neg)
  \item \code{block - optional} (Block number. If column is missing, it is calculated automatically by the "blockNum" argument.)
}}

\item{blockNum}{The number of blocks, with default value of max(data$block).}

\item{algoNum}{The chosen algorithm, with default value of 1.}
}
\value{
The results include EPT scores for each prime_cat, for each sid, overall and by block (for computing internal consistency).
}
\description{
Calculate EPT scores using one of the 9 algorithms recommended in Segal-Gordon et al. (2024).
Segal-Gordon et al. recommended Algorithms 1 and 7. Algorithm 1 is set as default.
}
\examples{
scores1 <- calc_ept_score(myProcessedData);
scores2 <- calc_ept_score(myProcessedData, algoNum = 7);
scores3 <- calc_ept_score(myProcessedDataWithoutBlockCol, blockNum = 5, algoNum = 3);
}
