% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_functions.R
\name{calc_ept_score}
\alias{calc_ept_score}
\title{EPT's G score calculation}
\usage{
calc_ept_score(data, algoNum = 1)
}
\arguments{
\item{data}{The processed data, with columns:
sid (subjects' unique id), 
error (1 is wrong, 0 is right), 
rt (reaction time), 
prime_cat (primes' categories), 
target_cat (pos / neg), 
block = part (EPT with 3 parts), 
trial - trial number.
The order should be: sid, error, rt, prime_cat, target_cat, block, trial, part.}

\item{algoNum}{The chosen algorithm, with default value of 1.}
}
\value{
Data frame of scores by subject (sid)
}
\description{
Calculate scores according to the users' choice of algorithm, with algorithm 1 as default
}
\examples{
scores1 <- calc_ept_score(myProcessedData);
scores2 <- calc_ept_score(myProcessedData, 7);
}
